update 16/03 comment and order

+++ src/main/java/com/bkap/fruitshop/service/impl/CommentServiceImpl.java
@@ -6,6 +6,8 @@ import com.bkap.fruitshop.entity.Comment;
 import com.bkap.fruitshop.exception.AppException;
 import com.bkap.fruitshop.exception.ErrorCode;
 import com.bkap.fruitshop.repository.CommentRepository;
+import com.bkap.fruitshop.repository.PostRepository;
+import com.bkap.fruitshop.repository.UserRepository;
 import com.bkap.fruitshop.service.CommentService;
 import lombok.RequiredArgsConstructor;
 import org.modelmapper.ModelMapper;
@@ -20,11 +22,17 @@ public class CommentServiceImpl implements CommentService {
 
     private final CommentRepository commentRepository;
     private final ModelMapper modelMapper;
+    private final UserRepository userRepository;
+    private final PostRepository postRepository;
 
     @Override
     public List<CommentResponse> findAllComments() {
         return commentRepository.findAll().stream()
-                .map((element) -> modelMapper.map(element, CommentResponse.class))
+                .map(comment -> {
+                    CommentResponse response = modelMapper.map(comment, CommentResponse.class);
+                    response.setUsername(comment.getUser().getUsername());
+                    return response;
+                })
                 .collect(Collectors.toList());
     }
 
@@ -32,14 +40,31 @@ public class CommentServiceImpl implements CommentService {
     public CommentResponse findCommentById(long id) {
         Comment comment =  commentRepository.findById(id)
                 .orElseThrow(() -> new AppException(ErrorCode.COMMENT_NOT_FOUND));
-        return modelMapper.map(comment, CommentResponse.class);
+
+        return mapToResponse(comment);
     }
 
     @Override
     public CommentResponse create(CommentRequest request) {
-        Comment comment = modelMapper.map(request, Comment.class);
-        Comment savedComment = commentRepository.save(comment);
-        return modelMapper.map(savedComment, CommentResponse.class);
+        var user = userRepository.findById(request.getUserId())
+                .orElseThrow(() -> new AppException(ErrorCode.USER_NOT_FOUND));
+
+        var post = postRepository.findById(request.getPostId())
+                .orElseThrow(() -> new AppException(ErrorCode.POST_NOT_FOUND));
+
+        Comment comment = new Comment();
+        comment.setUser(user);
+        comment.setPost(post);
+        comment.setContent(request.getContent());
+
+        if (request.getParentCommentId() != null) {
+            Comment parentComment = commentRepository.findById(request.getParentCommentId())
+                    .orElseThrow(() -> new AppException(ErrorCode.COMMENT_NOT_FOUND));
+            comment.setParentComment(parentComment);
+        }
+        comment = commentRepository.save(comment);
+        // Tạo response
+        return mapToResponse(comment);
     }
 
     @Override
@@ -47,12 +72,16 @@ public class CommentServiceImpl implements CommentService {
         Comment comment = commentRepository.findById(id)
                 .orElseThrow(() -> new AppException(ErrorCode.COMMENT_NOT_FOUND));
 
-        modelMapper.map(request, comment);
-        return modelMapper.map(commentRepository.save(comment), CommentResponse.class);
+        comment.setContent(request.getContent());
+
+        return mapToResponse(comment);
     }
 
     @Override
     public void delete(long id) {
+        if (!commentRepository.existsById(id)) {
+            throw new AppException(ErrorCode.COMMENT_NOT_FOUND);
+        }
         commentRepository.deleteById(id);
     }
 
@@ -64,7 +93,26 @@ public class CommentServiceImpl implements CommentService {
     }
 
     @Override
+    public List<CommentResponse> findAllCommentsByUserIdAndPostId(Long userId, long postId) {
+        return commentRepository.findByUserIdAndPostId(userId, postId).stream()
+                .map((element) -> modelMapper.map(element, CommentResponse.class))
+                .collect(Collectors.toList());
+    }
+
+    @Override
     public long countCommentsByPostId(long postId) {
         return commentRepository.countByPostId(postId);
     }
+
+    private CommentResponse mapToResponse(Comment comment) {
+        return CommentResponse.builder()
+                .id(comment.getId())
+                .parentCommentId(comment.getParentComment() != null ? comment.getParentComment().getId() : null)
+                .postId(comment.getPost().getId())
+                .content(comment.getContent())
+                .username(comment.getUser().getUsername()) // Đảm bảo ánh xạ đúng
+                .postTitle(comment.getPost().getTitle())
+                .build();
+    }
+
 }
